+++ ExEFIS/ExEFIS.vcxproj
@@ -77,6 +77,7 @@
     <ClCompile Include="adhrs.cpp" />
     <ClCompile Include="airspeed.cpp" />
     <ClCompile Include="altitude.cpp" />
+    <ClCompile Include="candriver.cpp" />
     <ClCompile Include="directional_gyro.cpp" />
     <ClCompile Include="ExEFIS.cpp" />
     <ClCompile Include="horizon_instrument.cpp" />
@@ -101,6 +102,7 @@
     <ClInclude Include="airspeed.h" />
     <ClInclude Include="altitude.h" />
     <ClInclude Include="ByteTransfer.h" />
+    <ClInclude Include="candriver.h" />
     <ClInclude Include="directional_gyro.h" />
     <ClInclude Include="horizon_instrument.h" />
     <ClInclude Include="HRS_9250.h" />
