+++ ExEFIS/candriver.h
@@ -1,0 +1,8 @@
+#pragma once
+class candriver
+{
+public:
+	candriver();
+	~candriver();
+};
+
